separate out filter and filter-init
