+++ crates/fluvio-smartmodule/examples/filter_init/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "fluvio-smartmodule-filter"
+name = "fluvio-smartmodule-filter-init"
 version = "0.0.0"
 authors = ["Fluvio Contributors <team@fluvio.io>"]
 edition = "2021"
@@ -9,4 +9,5 @@ publish = false
 crate-type = ['cdylib']
 
 [dependencies]
+once_cell = "1.13.0"
 fluvio-smartmodule = { path = "../../" }
